requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
captcha==0.4
crispy-bootstrap4==2022.1
dj-database-url==2.0.0
dj-static==0.0.6
Django
django-bootstrap-datepicker-plus==5.0.3
django-bootstrap4==23.1
django-crispy-forms==2.0
django-environ==0.10.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
gunicorn==20.1.0
Pillow==9.4.0
psycopg2==2.9.6
pydantic==1.10.8
soupsieve==2.4.1
sqlparse==0.4.3
static3==0.7.0
typing_extensions==4.6.2
tzdata==2022.7
